Refactor Wheel Of Fortune code into a subcomponent
